Refactoring major parts (currently breaks)
