+++ src/keyboard/keyboard_manager.h
@@ -1,0 +1,10 @@
+#include "roboteam_msgs/RobotCommand.h"
+
+class KeyboardManager {
+public:
+  KeyboardManager();
+  void handleSDLEvents();
+
+private:
+  int currentId = 0;
+}
